+++ src-ui/app/models/settings.ts
@@ -25,10 +25,10 @@ export interface AppSettings {
   hotkeys: { [hotkeyId: string]: string[] };
   hideSnowverlay: boolean;
   oscServerEnabled: boolean;
-  valveIndexMaxBrightness: number;
-  bigscreenBeyondMaxBrightness: number;
-  bigscreenBeyondUnsafeBrightness: boolean;
-  bigscreenBeyondForceFanBrightnessSafety: boolean;
+  valveIndexMaxBrightness: number; // User limit
+  bigscreenBeyondMaxBrightness: number; // User limit
+  bigscreenBeyondUnsafeBrightness: boolean; // Allow brightness above 150%
+  bigscreenBeyondBrightnessFanSafety: boolean; // Force fan to 100% if brightness is above 100%
 }
 
 export type QuitWithSteamVRMode = 'DISABLED' | 'IMMEDIATELY' | 'AFTERDELAY';
@@ -81,7 +81,7 @@ export const APP_SETTINGS_DEFAULT: AppSettings = {
   valveIndexMaxBrightness: 160,
   bigscreenBeyondMaxBrightness: 150,
   bigscreenBeyondUnsafeBrightness: false,
-  bigscreenBeyondForceFanBrightnessSafety: true,
+  bigscreenBeyondBrightnessFanSafety: true,
 };
 
 export type ExecutableReferenceStatus =