+++ src-ui/app/services/brightness-control/hardware-brightness-drivers/bigscreen-beyond-hardware-brightness-control-driver.ts
@@ -64,7 +64,6 @@ export class BigscreenBeyondHardwareBrightnessControlDriver extends HardwareBrig
     const hwValue = this.swValueToHWValue(hwPercentage);
     invoke('bigscreen_beyond_set_brightness', {
       brightness: hwValue,
-      fanSafety: settings.bigscreenBeyondForceFanBrightnessSafety,
     });
   }