+++ src-ui/app/views/dashboard-view/views/hmd-automations-view/tabs/hmd-automations-bigscreen-beyond-tab/hmd-automations-bigscreen-beyond-tab.component.ts
@@ -122,8 +122,7 @@ export class HmdAutomationsBigscreenBeyondTabComponent implements OnInit {
 
   toggleBigscreenBeyondForceFanBrightnessSafety() {
     this.appSettingsService.updateSettings({
-      bigscreenBeyondForceFanBrightnessSafety:
-        !this.appSettings.bigscreenBeyondForceFanBrightnessSafety,
+      bigscreenBeyondBrightnessFanSafety: !this.appSettings.bigscreenBeyondBrightnessFanSafety,
     });
     // Set brightness to same value to reset fan safety if needed
     this.hardwareBrightness.setBrightness(